+++ apps/expo/src/app/index.tsx
@@ -1,15 +1,9 @@
 import React, { useEffect, useState } from "react";
-import { Button, Text, View } from "react-native";
-import { RootSiblingParent } from "react-native-root-siblings";
+import { Text, View } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
 import * as Notifications from "expo-notifications";
 import { Stack } from "expo-router";
-import { SignedIn, SignedOut, useAuth } from "@clerk/clerk-expo";
 
-import ResetPassword from "~/components/ResetPassword";
-import SignInScreen from "~/components/SignInScreen";
-import SignInWithOAuth from "~/components/SignInWithOAuth";
-import SignUpScreen from "~/components/SignUpScreen";
 import registerForPushNotificationsAsync from "~/notifications/registerNotifications";
 
 Notifications.setNotificationHandler({
@@ -20,23 +14,6 @@ Notifications.setNotificationHandler({
       shouldSetBadge: false,
     }),
 });
-
-const SignOut = () => {
-  const { isLoaded, signOut } = useAuth();
-  if (!isLoaded) {
-    return null;
-  }
-  return (
-    <View>
-      <Button
-        title="Sign Out"
-        onPress={async () => {
-          await signOut();
-        }}
-      />
-    </View>
-  );
-};
 
 const Index = () => {
   const [, setExpoPushToken] = useState<Notifications.ExpoPushToken | null>(
@@ -50,25 +27,13 @@ const Index = () => {
   }, []);
 
   return (
-    <RootSiblingParent>
-      <SafeAreaView className="">
-        {/* Changes page title visible on the header */}
-        <Stack.Screen options={{ title: "Home Page" }} />
-        <View className="h-full w-full p-4">
-          <Text className="font-poppins400"> Index </Text>
-          <SignedIn>
-            <Text>You are signed in!</Text>
-            <SignOut />
-          </SignedIn>
-          <SignedOut>
-            <SignUpScreen />
-            <SignInScreen />
-            <SignInWithOAuth />
-            <ResetPassword />
-          </SignedOut>
-        </View>
-      </SafeAreaView>
-    </RootSiblingParent>
+    <SafeAreaView className="">
+      {/* Changes page title visible on the header */}
+      <Stack.Screen options={{ title: "Home Page" }} />
+      <View className="h-full w-full p-4">
+        <Text className="font-poppins400"> Index </Text>
+      </View>
+    </SafeAreaView>
   );
 };
