+++ apps/expo/src/hooks/warmUpBrowser.tsx
@@ -1,8 +1,8 @@
-import React from "react";
+import { useEffect } from "react";
 import * as WebBrowser from "expo-web-browser";
 
 export const useWarmUpBrowser = () => {
-  React.useEffect(() => {
+  useEffect(() => {
     void WebBrowser.warmUpAsync();
     return () => {
       void WebBrowser.coolDownAsync();
